fix overflow

+++ src/Smooth.h
@@ -15,8 +15,9 @@ typedef void (*FNcallback)(double const /* new_value */);
 
 class Smooth {
     private:
-    int set_size;
-    int count;
+    uint16_t set_size;
+    uint32_t count;
+
     FNcallback cbchange;
     FNcallback cblower;
     FNcallback cbupper;
@@ -38,7 +39,7 @@ class Smooth {
     double get_avg() const;
 
     // get the total sample count
-    int get_count() const;
+    uint32_t get_count() const;
 
     // get the current window size (num samples)
     int get_window() const;
